requirements.txt
flask==3.0.0
google-api-python-client==2.108.0
python-dotenv==1.0.0
requests==2.31.0
isodate==0.6.1
